pyproject.toml
[project]
name = "agml-chat"
version = "0.1.0"
description = "nanochat-style multimodal fine-tuning + chat interface for AgML classification datasets"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "agml>=0.7.4",
  "accelerate>=0.34.0",
  "datasets>=3.0.0",
  "fastapi>=0.115.0",
  "numpy>=1.26.0",
  "peft>=0.12.0",
  "pillow>=10.0.0",
  "pydantic>=2.8.0",
  "python-multipart>=0.0.9",
  "pyyaml>=6.0.1",
  "transformers>=4.45.0",
  "uvicorn>=0.30.0",
]

[dependency-groups]
dev = [
  "ipykernel>=7.1.0",
  "matplotlib>=3.10.0",
  "python-dotenv>=1.0.1",
  "pytest>=8.0.0",
  "ruff>=0.6.0",
]

[project.optional-dependencies]
cpu = [
  "torch>=2.2.0",
  "torchvision",
]
gpu = [
  "torch>=2.2.0",
  "torchvision",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agml_chat", "scripts"]

[tool.setuptools.package-data]
agml_chat = ["templates/*.html"]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "gpu" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "gpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv]
default-groups = []
conflicts = [
  [
    { extra = "cpu" },
    { extra = "gpu" },
  ],
]
